workflows/index-generation/requirements.txt
pandas==1.4.3
marisa-trie==0.7.7
sourmash==4.8.2
biopython==1.81
